api/requirements.txt
attrs==22.2.0
autopep8==2.0.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.38.0
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.23.5
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pycodestyle==2.10.0
PyMySQL==1.0.3
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
uWSGI==2.0.21
Werkzeug==2.2.3
